moved metrics battery
