+++ src/metrics/batteries.rs
